backend/leavereqpro/requirements.txt
Django==2.2.5
django-rest-auth==0.9.5
djangorestframework==3.10.3

